// src/models/cliente.new.dto.ts
export interface ClienteNewDTO {
    nome: string;
    email: string;
    cpfOuCnpj: string;
    tipo: string;
    senha: string;
    logradouro: string;
    numero: string;
    complemento: string;
    bairro: string;
    cep: string;
    telefone1: string;
    telefone2: string;
    telefone3: string;
    cidadeId: string;
}